issue83 Move LRAInfo into TCK
